@@ -5,7 +5,7 @@ import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 import {UserProfileComponent} from "./components/user-profile/user-profile.component";
 import {MaskEmailPipe} from "./helpers/pipe/mask-email.pipe";
-import {HttpClientModule} from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import {DataListComponent} from './components/data-list/data-list.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {PostListComponent} from './components/post-list/post-list.component';
@@ -24,37 +24,30 @@ import {MatCardModule} from "@angular/material/card";
 import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
 import {MatAutocompleteModule} from "@angular/material/autocomplete";
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    UserProfileComponent,
-    DataListComponent,
-    PostListComponent,
-    HomeComponent,
-    AboutComponent,
-    ContactComponent,
-    CalcComponent,
-    NavBarComponent,
-    WeatherComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    MaskEmailPipe,
-    HttpClientModule,
-    FormsModule,
-    BrowserAnimationsModule,
-    MatButtonModule,
-    MatIconModule,
-    MatInputModule,
-    MatSelectModule,
-    ReactiveFormsModule,
-    MatCardModule,
-    MatProgressSpinnerModule,
-    MatAutocompleteModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        UserProfileComponent,
+        DataListComponent,
+        PostListComponent,
+        HomeComponent,
+        AboutComponent,
+        ContactComponent,
+        CalcComponent,
+        NavBarComponent,
+        WeatherComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        MaskEmailPipe,
+        FormsModule,
+        BrowserAnimationsModule,
+        MatButtonModule,
+        MatIconModule,
+        MatInputModule,
+        MatSelectModule,
+        ReactiveFormsModule,
+        MatCardModule,
+        MatProgressSpinnerModule,
+        MatAutocompleteModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {
 }
